Decouple service logic from the storage logic
